feature: add nested transition example
