+++ demos/src/Examples/Transition/Vue/VueComponent.vue
