+++ packages/vue-3/src/useEditor.ts
@@ -1,5 +1,5 @@
 import { EditorOptions } from '@tiptap/core'
-import { onBeforeUnmount, onMounted, shallowRef } from 'vue'
+import { onMounted, shallowRef } from 'vue'
 
 import { Editor } from './Editor.js'
 
@@ -8,10 +8,6 @@ export const useEditor = (options: Partial<EditorOptions> = {}) => {
 
   onMounted(() => {
     editor.value = new Editor(options)
-  })
-
-  onBeforeUnmount(() => {
-    editor.value?.destroy()
   })
 
   return editor
